pyproject.toml
[tool.poetry]
name = "blogapi"
version = "0.1.0"
description = ""
authors = ["IIH <ds5105119@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amqp = "5.2.0"
asgiref = "3.8.1"
attrs = "24.2.0"
billiard = "4.2.0"
black = "24.8.0"
boto3 = "*"
botocore = "*"
celery = "5.4.0"
certifi = "*"
cffi = "*"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cmdstanpy = "1.2.4"
contourpy = "1.3.0"
cryptography = "*"
cycler = "0.12.1"
defusedxml = "0.7.1"
dj-rest-auth = "6.0.0"
django = "*"
django-allauth = "64.2.1"
django-cors-headers = "4.4.0"
django-elasticsearch-dsl = "8.0"
django-environ = "0.11.2"
django-filter = "24.3"
django-nine = "0.2.7"
django-storages = "1.14.4"
django-taggit = "6.0.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "*"
elastic-transport = "*"
elasticsearch = "*"
elasticsearch-dsl = "*"
fonttools = "4.53.1"
gunicorn = "23.0.0"
h11 = "0.14.0"
holidays = "0.56"
idna = "3.8"
importlib-resources = "*"
inflection = "0.5.1"
jmespath = "1.0.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
kiwisolver = "*"
kombu = "5.4.0"
matplotlib = "3.9.2"
mypy-extensions = "1.0.0"
numpy = "*"
oauthlib = "3.2.2"
packaging = "24.1"
pandas = "2.2.2"
pathspec = "0.12.1"
pillow = "10.4.0"
platformdirs = "4.3.2"
prompt-toolkit = "3.0.47"
prophet = "1.1.5"
pycparser = "2.22"
pyjwt = "2.9.0"
pyparsing = "*"
python-dateutil = "2.9.0.post0"
python3-openid = "3.2.0"
pytz = "2024.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rpds-py = "0.20.0"
s3transfer = "0.10.2"
setuptools = "*"
six = "1.16.0"
sqlparse = "0.5.1"
stanio = "0.5.1"
tqdm = "4.66.5"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.2"
uvicorn = "0.30.6"
vine = "5.1.0"
wcwidth = "0.2.13"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
